Fix timing attack in mgk_memeql
mgk_memeql was vulnerable to a subtle timing attack disclosing data that was
being compared. In particular, a[i] == b[i] would incur a processor branch and
may disclose whether or not the branch was taken (one extra jne/je instruction
on x86, for example), see [1]. Fixed to do a constant-time bytewise xor that
hopefully does not leak any data.

[1] http://rdist.root.org/2010/01/07/timing-independent-array-comparison/