Modify Yami to always tunnel
The Yami/Yugi infrastructure has so far worked in such a way that before the
handshake was finished, Yami was requiring packets from Yugi in a size-only
fashion. This was severely limiting the flexibility of the system, in that the
initial client message had to be of a fixed width, which impairs implementation
of Megaki RESTART packet. This change removes Yami's expected-length response
parameter and Yugi now passes a message header to Yami for evaluation, whereas
an expected length is then derived from the header. This allows not only
greater flexibility for Megaki, but also for implementation of any strict
request-response protocol where packet length can be derived from a
fixed-length header (as it ought to be).