Formalized first version of the Sazukari API
Created a simple API for Sazukari client library. Since we need to call the
functions from the NDK, I've attempted to minimize any outer-world dependency
we might have (we can't get rid of libcrypto, though). The current version of
the API is small and self-explanatory and should stay this way; I see no reason
to complicate it beyond this.