Swapped pthread mutex with a POSIX semaphore
So far, conn_t::recvmut was a pthread_mutex_t. This has worked fine so far, but
we require that one thread locks it (main thread loop) and another one (a
worker) unlocks it so the buffer cannot be overwritten with more data while a
worker thread has not yet acquired the job. Mutex by definition cause undefined
behavior in this instance, so I had to opt for POSIX semaphores instead.