Rectified mutex (semaphore) lock/unlock issue
There was a subtle mistake that allowed on_data_read for a particular
connection to be called while a threadpool has not yet acquired a
job_handle_message job. In a carefully crafted setting, this would allow a
connection to be completely terminated while another thread is ignorant of this
fact and processes the received message. It is unclear to me if this is even a
real problem, but a it's better for a condition like this to be eliminated as
soon as possible.