Define new Sazukari API
I have amended Sazukari files to include a very different API approach. Instead
of operating on a state machine and being notified of inbound data (like Yami),
Sazukari now takes recv()- and send()-like callbacks and has big methods like
szkr_do_handshake which run the full procedure and bail out if something goes
wrong. This way, the state is implicitly the context inside the function: much
easier to maintain. NB that this does not impair usability or trashes something
important in favor of simplicity: since Sazukari is a client library, it makes
the most sense to block on I/O (we can run this from a separate thread on the
device, for example) as we will *absolutely* never need mass client
connections. No actual library code has been written, tho: the changes are only
in the headers so far.