Implementation of RSA keygen and housekeeping
Sazukari internal function to do a RSA keygen has been implemented (but not
tested). A public API change now ensures that a server public key in serialized
form can be passed to Sazukari. Also a small change to common.h and megaki.h
has been introduced, to allow some convenience macros to be used in sazukari.h.
Changes will be pulled into mkd directories and I expect to create a common
folder for that type of thing.