Fix fatal bug in timeout handling in Yugi
Yugi had a bug when handling timeouts which resulted in a zombie libuv
timer being summoned from an on_write callback. This manifested itself
when the server was hammered with a rapid succession of SYN packets: it
would kill the timeout timer but it would be in turn restarted when the
packet finishes writing. This commit hopefully fixes the bug. I am
saying "hopefully" because as far as I can see no such condition can
happen again. Additionally, the server has been tested with a simple
stress test (perennial_quest.bash) for 100 minutes straight and no
failures occurred.

The result of the bug occurring were a failed assertion in libuv source
code in normal use case (but actually that code accessed free'd memory
so that was undefind behavior) or a segfault when closing a particular
connection (if the free'd memory was subsequently overwritten).