Add more debug info and foolproof errcheck
Yami now outputs diagnostic info when it starts and finishes assembling
a SYN-ACK packet. It also checks for errors on RAND_bytes and
RAND_pseudo_bytes in a manner that catches all errors that libcrypto may
encounter during these. (Before, if they hypothetically return -1, Yami
happily continues and doesn't notice it has no random data in the
buffers, this can be fatal.)