Implement ACK assembling
Sazukari can now assemble an ACK packet and establish a shared secret
over a fully insecure channel, provided it has the right server cert on
file. Of course, AFAICT, a man-in-the-middle would be powerless and
could only perform a denial-of-service. No confidential data could be
extracted even if someone in the middle could inject packets at will.
Also, compromise of a server certificate results in no compromise of
previous conversations, provided that a client RSA key was not
compromised. (This would be very hard to perform as the client RSA is
temporary and only stored in memory, but if someone could extract data
from memory, they would have a lot more powerful attacks to perform.)

The next step is implementation of a session-restart packet which would
allow a client to reestablish a connection that was terminated earlier,
in order to avoid regeneration of its RSA key. However, the server is
not obliged to allow this and may prompt the client to perform a full
rehandshake if necessary.