Add support for message sending
Sazukari is now able to send arbitrary messages to the mkd daemon (the
daemon is tested and it can successfully decrypt them). However, as of
now, Sazukari doesn't know how to receive the reply and the server
doesn't know how to send one, this will be added later. For now, it
works as it is.