Fix problem with HMAC
Yami and Sazukari now do an HMAC of both the IV and the ciphertext, as
opposed to only ciphertext. This was a major security risk as it allowed
an adversary to modify the IV without the change being detected. As CBC
is partly malleable, this could mangle a ciphertext block but allow for
arbitrary bit flipping in the one following it.