Fix ignored assertions when debugging
I come from a gamedev background where every cycle counts and you can't
really afford to waste them on assertions. Gamedev is really just for
fun, so if assertions fail it's never a big deal. (Unless we're dealing
with a networking module, but that's another story.) As such, it was a
reflex for me to disable assertions (replace them with dummy macros)
when not debugging, but it's a huge mistake. Huge, huge, huge mistake.
'Cause in games, if a bug arises because some assertion has failed to
catch it, the worst that can happen is that people notice it and upload
Youtube videos of it (given that your game is popular enough). Even if
assertions catch such bugs, people will be pissed off, so it's better to
just ignore them. Games are self-contained, do not depend on external
input besides the mouse and keyboard, so such mistakes never attribute
to anything. But here it's not the case. Here we deal with crypto,
security and the like, we simply CANNOT in our right minds decide to
ignore assertions like I just did. So I apologize for being a retard. I
have since fixed my ways.