Integrate Pegasus. Fix bugs with Yugi
Some small bugs in Yugi have been fixed and Pegasus has been
kinda-integrated with Yami (although a long way to go still). There is
no functional minion code so this currently is an obstacle to more
thorough testing. Some sample code will have to be written in order to
facilitate proper unit-testing of Pegasus---this will probably mean
actual work on the Arcangelo surrogate (PHP). For now, I've tested that
this compiles and attempts to spin up a minion after a successful
connection but the minion is a dummy infinite sleep function so it never
responds, Pegasus deems it unstable and then just restarts the process.
I haven't stress-tested this, though, so there may be some race
conditions or such that can be revealed only with way more strict tests.