Add preliminary message passing functionality
Yami can pass a received, decrypted message to Pegasus and Pegasus can
pipe this to a broker which can reply. (Yami still cannot encrypt and
send the response out, though.) Also, Yugi now passes a payload to Yami
and this gets passed all the way to the broker so it knows some of the
context inside it.

More importantly, we have switched from a broker-per-connection to a
broker-per-request model. Since we aim for a REST alternative with the
Yami/Yugi/Pegasus/Arcangelo/PHP stack, there should not be
connectionwide context, which allows us to do broker-per-request. And
this is way better: we prefork brokers and if there are lot of
timing-out connections, the brokers would just sleep, waiting for new
messages from the clients to arrive, which would quickly deplete all of
them. When it's like this, however, a new broker is allocated for each
request and then deallocated as soon as it finishes, this will hopefully
enable better parallelization too. There are a few (unwanted) debug log
outputs which will be removed later, when the full functionality is
available.