Offload message encryption to commons
Since both the client and the server need to encrypt and decrypt
messages, it makes sense to move this code to the common libraries. In
this commit, encryption has been moved to common.c and libsazukari was
amended to use this code.

OpenSSL code is the ugliest piece of fuck there ever was. I should have
really used NSS or something for this code. This is ridiculous.