Implement growing buffer for big messages in Yugi
Yugi can now support rather big message sizes without allocating great
amounts of memory if it isn't necessary. In particular, it can now have
a dynamic receive buffer which is used if the message exceeds a (low)
fixed amount, and the buffer grows as new data is received. For the
first time, a single message could be spread across multiple
on_data_read calls. However, as the whole message is kept in memory,
this is still not satisfactory for transferring any kind of streaming
data. For streaming to and from files, a new method will be implemented
which will add a new class of functions to the Yami interface and which
would allow Yugi to efficiently send out binary files (and receive them
to be uploaded) to clients.

A couple of bugs were also fixed, but these were not so critical, mainly
sanity checks and Valgrind warnings. Also, the compiler is spitting out
some warnings for the source files, so this will have to be rectified I
guess.

I'm not really confident that everything is stable here, tho: I wouldn't
be surprised if some double-free's or segfaults show up in contended
environments. I will be writing a very, very thorough unit testing
framework to test the whole stack (a Spartan job, if you will, but I am
really passionate about killing off hard-to-trace bugs such as race
conditions) and I'll be doing a full-blown code review later on.