Change to the Megaki protocol
A change was added into the core protocol (and implemented immediately
in both MKD and Sazukari) which would render ineffective a class of
theoretical attacks.

So far, if an adversary was to capture a client's public RSA key (meant
to be kept secret and strictly temporary, so such a capture should not
be possible), it could use this data to impersonate the server.
Obviously, any impersonation of the server without access to the
original certificate is a very serious flaw, but the RSA keys the client
uses are *never* reused and never stored onto any medium and only kept
in memory and regenerated as needed, so only a determined attacker with
access to a user's running phone would *in theory* be able to find out
these keys (and this defeats the purpose as such a key is effective only
inside that mobile device). However, if it happened in some way that I
am not aware is possible, that the attacker gets hold of these keys
(maybe exploiting the server in some way, for example?), it could then
simply impersonate the server. The only secret data sent by the client
to the server is the public key itself, and if an attacker knew this
beforehand, she would have the same knowledge about the client as the
server itself, without needing the server's RSA private key to decrypt
the initial packet.

With this change, however, the client transfers an ephemeral key
intended to be used only for the subsequent packets before a key
agreement is reached. The server is required to HMAC-authenticate its
SYNACK packet, and the client is required to HMAC-authenticate its ACK
packet then. This way, an adversary could not impersonate the server
even if it somehow got hold of the client's public key, as now it needs
the server's certificate to find out the ephemeral one needed to sign
messages afterwards.