Half-implement session resume handling
Yami can now handle the first MSG-RSTOR packet, but it does not know how
to respond. (And neither would libsazukari know how to handle it, so
we'll consider this one a tie.) Still WIP, people, but I hope I can
patch together full support for this today or maybe early tomorrow.